kis_cubic_curve.cpp: initialize variables
Fixes Coverity CID 709308: Uninitialized scalar field (UNINIT_CTOR)

Non-static class member ""m_begin"" is not initialized in this constructor nor in any functions that it calls.
Non-static class member ""m_end"" is not initialized in this constructor nor in any functions that it calls.
Non-static class member ""m_intervals"" is not initialized in this constructor nor in any functions that it calls.